server: onhover remove old code
