+++ language-server/src/features/hover/hover.ts
@@ -1,20 +1,14 @@
 import { inject, injectable } from 'tsyringe'
-import { Connection, MarkupContent } from 'vscode-languageserver'
+import { Connection } from 'vscode-languageserver'
 import { Provider } from '../provider'
 import * as winston from 'winston'
 import * as ls from 'vscode-languageserver'
-import { RangeConverter } from '../../utils/rangeConverter'
 import { URI } from 'vscode-uri'
 import { LoadFolder } from '../../mod/loadfolders'
 import { ProjectManager } from '../../projectManager'
 import { Project } from '../../project'
-import { Element } from '@rwxml/analyzer'
-import { AsEnumerable } from 'linq-es2015'
-import { TextDocumentManager } from '../../textDocumentManager'
-import { Definition } from '../definition'
-import { FileStore } from '../../fileStore'
-import { DependencyFile } from '../../fs'
 import { LogToken } from '../../log'
+import { ReferenceHover } from './reference'
 // how to use 'prettydiff' (it is quite different to use than other standard libs)
 // https://github.com/prettydiff/prettydiff/issues/176
 // https://github.com/sprity/sprity/blob/master/lib/style.js#L38-L53
@@ -62,10 +56,7 @@ export class HoverProvider extends Provider {
   constructor(
     loadFolder: LoadFolder,
     projectManager: ProjectManager,
-    private readonly rangeConverter: RangeConverter,
-    private readonly textDocumentManager: TextDocumentManager,
-    private readonly definitionProvider: Definition,
-    private readonly fileStore: FileStore,
+    private readonly refHover: ReferenceHover,
     @inject(LogToken) baseLogger: winston.Logger
   ) {
     super(loadFolder, projectManager)
@@ -87,102 +78,13 @@ export class HoverProvider extends Provider {
   }
 
   private onHover(projects: Project[], uri: URI, position: ls.Position): ls.Hover | null | undefined {
-    const text = this.textDocumentManager.get(uri.toString())?.getText()
-    if (!text) {
-      return
+    for (const proj of projects) {
+      const res = this.refHover.onReferenceHover(proj, uri, position)
+      if (res) {
+        return res
+      }
     }
 
-    const defs = projects.map((proj) => this.definitionProvider.findDefs(proj, uri, position)).flat()
-    if (defs.length === 0) {
-      return
-    }
-
-    const def = defs[0]
-
-    const contents: MarkupContent = {
-      kind: 'markdown',
-      value: '',
-    }
-
-    const formattedXML = this.getTargetXMLString(def)
-    const sourceUri = URI.parse(def.document.uri)
-    const sourceFile = this.fileStore.get(def.document.uri)
-    let packageId = 'local'
-    if (sourceFile && DependencyFile.is(sourceFile)) {
-      packageId = sourceFile.ownerPackageId
-    }
-
-    // https://code.visualstudio.com/api/extension-guides/command#command-uris
-    // https://code.visualstudio.com/api/references/commands
-
-    contents.value += `\
-\`\`\`xml
-${formattedXML}
-\`\`\`
--------
-packageId: \`${packageId}\`  
-source: [${sourceUri.fsPath}](${sourceUri.toString()})
-`
-
-    // NOTE: property range 는 뭐하는거지?
-    return { contents }
-  }
-
-  /**
-   * getTargetXMLString returns minified version of the target xml def
-   */
-  private getTargetXMLString(elem: Element): string {
-    // find defName node
-    const defNameNodeIndex = elem.ChildElementNodes.findIndex((el) => el.tagName === 'defName')
-    if (defNameNodeIndex >= elem.ChildElementNodes.length) {
-      // if no defName node exists?
-      return 'GET_TARGET_XML_STRING_UNDEFINED'
-    }
-
-    // pick other nodes to show.
-    // does it returns null if take is over array length?
-    const total = 3
-    const childNodes = AsEnumerable(elem.ChildElementNodes).Skip(defNameNodeIndex).Take(total).ToArray()
-
-    if (childNodes.length < total) {
-      const take = defNameNodeIndex - 1
-      const nodes2 = AsEnumerable(elem.ChildElementNodes)
-        .Take(take >= 0 ? take : 0)
-        .Reverse()
-        .Take(total - childNodes.length)
-        .ToArray()
-
-      childNodes.push(...nodes2)
-    }
-
-    // stringify nodes and minify
-    const clonedDef = elem.cloneNode()
-    clonedDef.children = childNodes // does it breaks capsulation?
-    const raw = clonedDef.toString()
-
-    // format using prettydiff
-    let formatted = ''
-    try {
-      // https://github.com/prettydiff/prettydiff/blob/master/index.d.ts
-      // https://stackoverflow.com/questions/19822460/pretty-diff-usage/30648547
-      // https://github.com/prettydiff/prettydiff/tree/101.0.0#nodejs
-      prettydiff.options.source = raw
-      formatted = prettydiff()
-    } catch (err) {
-      console.error((err as Error).message)
-    }
-
-    const lines = formatted.split('\n')
-    const i = 7
-    if (lines.length > i) {
-      formatted = lines.slice(0, i).join('\n')
-      const indent = lines[i - 1].match(/\B( )+/)?.[0] as string
-      formatted += '\n' + indent + '...'
-      formatted += '\n' + lines[lines.length - 1]
-    } else {
-      formatted = lines.join('\n')
-    }
-
-    return formatted
+    return null
   }
 }
